requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic-settings>=2.0
PyJWT>=2.8
prometheus-fastapi-instrumentator>=7.0
pytest>=8.0
httpx>=0.27
requests>=2.32

